Introduce modules and update valkeyrie to v1.0.0
This commit introduces go modules to the leadership library.
It also updates the valkeyrie key value store abstraction
library to v1.0.0 which contains various patches and fixes
as well as a new modular store setup with separate repositories
for each supported drivers.

Since this new setup removed the use of a mock store, we
reintroduce it here as to keep the tests intact and avoid
the need to setup and initialize a store to test the core
logic of leader/follower/elections.

The behavior should remain intact with this change.

Signed-off-by: Alexandre Beslic <ab21c8b9f7@abronan.com>